==> 1_code/rNotebooks/1_setup/1_bird_data.Rmd <==

```{r setup, include=FALSE, cache=FALSE}
#Set root directory to R project root
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
```

 

```{css, echo=FALSE}
# set max height of code chunks using the following css styling

pre {
  max-height: 300px;
  overflow-y: auto;
}

pre[class] {
  max-height: 500px;
}
```



```{r message=FALSE}
library(dplyr)
library(stringr)
library(tidyr)
library(kableExtra)
library(plyr)
library(sf)
library(tmap)
library(tmaptools)
library(OpenStreetMap)
library(lubridate)
```


### Load data {-}

Data is from the boreal avian modelling projects point count database [REF]

```{r readBAM, eval=FALSE}
BAM<-read.csv("../../data/avian_lifeHistory/BAM_avianDatabase/tables/BAM_ss_spp.csv", header = TRUE)
#make copy
b<-BAM 

# rename columns
names(b)[7]<- "x"
names(b)[8]<- "y"

#Change myrtle warbler to yellow-rumped warbler
b$SPECIES[b$SPECIES == "MYWA"] <- "YRWA"
```


### Filter Calling Lake project {-}

```{r filterCL, eval=FALSE}
bird_callingLake_long<-b%>%
    filter(str_detect(SS, "^CL"))

#add a year column
library(lubridate)
bird_callingLake_long$DATE<-as.Date(bird_callingLake_long$DATE)
bird_callingLake_long$year<-year(bird_callingLake_long$DATE)


save(bird_callingLake_long, file="0_data/external/bird_callingLake_long.rData")
```

```{r echo=FALSE, message = FALSE, results="asis"}
load("0_data/external/bird_callingLake_long.rData")

knitr::kable(head(bird_callingLake_long)) %>%
  kable_styling(font_size = 10, bootstrap_options= c("striped", "hover", "condensed" ), full_width= F, position="center") %>%
  column_spec(1, width= "20em")%>%
  scroll_box(width = "100%", height = "400px")
```

### Filter birds {-}

Subset data frame so it does not include non-bird defections (e.g. wind, traffic, etc.) and unknown species. 

```{r eval=FALSE}
load("0_data/external/bird_callingLake_long.rData")

#Load bird alpha codes and corresponding species names.
bird_alpha<-read.csv("../../data/avian_lifeHistory/IBP-Alpha-Codes20.csv")


#Get a list of species that don't match alpha codes
noMatch<-anti_join(bird_callingLake_long, bird_alpha, by=c("SPECIES"="SPEC"))
unique(noMatch$SPECIES)

#Fix alpha codes
bird_callingLake_long["SPECIES"][bird_callingLake_long["SPECIES"] == "GRAJ"] <- "CAJA"
bird_callingLake_long["SPECIES"][bird_callingLake_long["SPECIES"] == "Oven"] <- "OVEN"

#Filter detection to only include known bird species.
bird_callingLake_long_sub<-semi_join(bird_callingLake_long, bird_alpha, by=c("SPECIES"="SPEC"))

#remove unidentified birds
bird_callingLake_long_sub<-bird_callingLake_long_sub[!grepl("^UN", bird_callingLake_long_sub$SPECIES),]

save(bird_callingLake_long_sub, file="0_data/manual/bird/bird_callingLake_long_sub.rData")

```


### Format data {-}

#### Convert bird data to wide format {-}

```{r eval=FALSE}
#Convert bird data to wide presence/absence format
# replace abundance values with 1
bird_callingLake_long_sub_occ<-bird_callingLake_long_sub
bird_callingLake_long_sub_occ$ABUND<-1
bird_callingLake_long_sub_occ<-distinct(bird_callingLake_long_sub_occ)

bird_callingLake_wide_occ<-bird_callingLake_long_sub_occ%>%
  dplyr::select(SS, PKEY, year, DATE, TIME, MAXDUR, MAXDIS, x, y, surveyType, SPECIES, ABUND)%>%
  tidyr::pivot_wider(names_from=SPECIES, values_from=ABUND, values_fill = 0)

save(bird_callingLake_wide_occ, file="0_data/manual/bird/bird_callingLake_wide_occ.rData")


bird_callingLake_long_sub<-distinct(bird_callingLake_long_sub)
bird_callingLake_wide_abund<-bird_callingLake_long_sub%>%
  dplyr::select(SS, PKEY, year, DATE, TIME, MAXDUR, MAXDIS, x, y, surveyType, SPECIES, ABUND)%>%
  distinct()%>%
  tidyr::pivot_wider(names_from=SPECIES, values_from=ABUND, values_fn={sum}, values_fill = 0)


save(bird_callingLake_wide_abund, file="0_data/manual/bird/bird_callingLake_wide_abund.rData")
```

```{r echo=FALSE, message = FALSE, results="asis"}
load("0_data/manual/bird/bird_callingLake_wide_abund.rData")

knitr::kable(head(bird_callingLake_wide_abund)) %>%
  kable_styling(font_size = 10, bootstrap_options= c("striped", "hover", "condensed" ), full_width= F, position="center") %>%
  column_spec(1, width= "20em")%>%
  scroll_box(width = "100%", height = "400px")
```

### Define and plot study area {-}

Create spatial object with point count locations.

```{r eval=FALSE}
ss_xy<-bird_callingLake_wide%>%
  dplyr::select(SS, x, y)%>%
  distinct()

ss_xy<-st_as_sf(ss_xy, coords=c("x","y"), crs=4326)

sx_xy<-st_transform(ss_xy, crs=3400)

save(ss_xy, file="0_data/manual/bird/ss_xy.rData")
```

Create bounding box around stations

```{r eval=FALSE}
#station bounding box
bb<-st_bbox(ss_xy)
save(bb, file="0_data/manual/bird/studyarea.rData")

# calling lake bounding box
cl_geo<-st_read("../../data/spatial/external/Alberta_Landcover_Data/Calling_Lake/surficialGeology/surf_ln_ll.shp")

cl_bb<-st_bbox(cl_geo)

#custom bounding box
library(rgeos)
# c_bb<-bbox2SP(55.35,55.15, -113.6, -113.15,bbox=NA,proj4string=st_crs("+init=epsg:3400"))

# c_bb_1<-st_bbox(c(xmin=55.35,xmax=55.15, ymin=-113.6, ymax=-113.15),crs=st_crs(ss_xy))
c_bb<-st_bbox(c(xmin=-113.6, ymin= 55.15, xmax=-113.15, ymax=55.35),crs=st_crs(ss_xy))


c_bb<-st_as_sfc(c_bb)
c_bb<-st_as_sf(c_bb)

save(c_bb, file="0_data/manual/bird/studyarea_big.rData")
```

Plot study area

```{r eval=FALSE}
#set up basemap
base<-read_osm(c_bb, type="bing", minNumTiles = 10)

#alberta boundary
alberta<-st_read("../../data/spatial/external/Alberta_Boundaries_and_Roads/Alberta/Alberta.shp")


mainmap<-tm_shape(base)+
  tm_rgb()+ # to color the satelite basemap
  tm_shape(ss_xy)+
    tm_symbols(shape = 4, alpha = 1, size = .6, col = "red")+
  tm_scale_bar(position=c("left", "BOTTOM"), text.color = "white", color.light="lightgrey")+
  tm_graticules(lines=FALSE)

insetmap<-tm_shape(alberta)+tm_fill(col="lightgrey")+tm_borders(lwd=.9, col="black")+
  tm_shape(c_bb)+tm_borders(lw=2, col="red") +
  tm_layout(inner.margins = c(0.04,0.04,0.04,0.04), outer.margins=c(0,0,0,0), bg.color="transparent", frame = FALSE)
    #tm_symbols(shape = 20, alpha = .5, border.col="dimgray", size = .1, col = "black")
  #tm_legend(position=c("left", "top"), frame=TRUE)

#Get aspect ratio of bounding box
c_bb_2<-st_bbox(c_bb)
asp <- (c_bb_2$ymax - c_bb_2$ymin)/(c_bb_2$xmax - c_bb_2$xmin)


library(grid)
w <- 1
h <- asp * w
vp <- viewport(0.86, 0.33, width = w, height=h)
#vp <- viewport(0.9, 0.22, width = w, height=h, just=c("right", "top"))

tmap_save(mainmap,filename="3_output/maps/studyArea_inset.png",
          dpi=300, insets_tm=insetmap, insets_vp=vp,
          height=asp*300, width=160, units="mm")

```

```{r studyArea_tmap, echo=FALSE, fig.cap= "Point count locations near Calling Lake, Alberta.", out.width = '100%'}
knitr::include_graphics("3_output/maps/studyArea_inset.png")
```



#### Filter data {-}

Filter data to within 10 years of the LiDAR acquistion

##### Import LiDAR aquistion shape file {-}

```{r eval=FALSE}
#Lidar acquistions
acqDate<-st_read("../../data/spatial/external/Alberta_Landcover_Data/LiDAR/GOA_LiDAR_Alberta/from_ChrisBater/lidar_acquisitions/acquisition_date.shp")

#stations
load("0_data/manual/bird/ss_xy.rData")

#study area bounding box
load("0_data/manual/bird/studyarea.rData")
bb<-st_as_sfc(bb)
bb<-st_as_sf(bb)

#Create a year column
acqDate$date-as.Date(acqDate$date)

# clip acquistions to study area

acqDate<-st_transform(acqDate, crs = st_crs(ss_xy))
acqDate_clip<-st_intersection(acqDate, bb)
save(acqDate_clip, file="0_data/manual/spatialCov/LiDAR/acqDate_clip.rData")

acqDate$year<-as.factor(year(acqDate$date))
m<-tm_shape(alberta)+tm_borders()+
  tm_shape(acqDate)+
  tm_fill(col="year", palette = "PuRd", contrast = c(0.26, 1), title="LiDAR acquisition year")+
  #tm_polygons(col=NA, border.col="black")+
  tm_layout(frame=FALSE)+
  tm_legend(outside=TRUE)

tmap_save(m, "3_output/maps/LiDAR_acqYear_Alberta.png", asp=0, height = 5, width = 6, outer.margins=c(0,0,0,.08))

acqDate_clip$year<-as.factor(year(acqDate_clip$date))
m<-tm_shape(acqDate_clip)+tm_borders()+
    tm_fill(col="year", palette = "PuRd", contrast = c(0.5, .6), title="LiDAR acquisition year")+
  tm_shape(ss_xy)+
    tm_symbols(shape = 4, alpha = 1, size = .6, col = "black")+
  tm_layout(frame=FALSE)+
  tm_legend(outside=TRUE)

tmap_save(m, "3_output/maps/LiDAR_acqYear_CallingLake.png", asp=0, height = 5, width = 8)


```

```{r echo=FALSE, fig.cap= "LiDAR acquisition dates, Alberta.", out.width = '100%'}
knitr::include_graphics("3_output/maps/LiDAR_acqYear_Alberta.png")
```


```{r echo=FALSE, fig.cap= "LiDAR acquisition dates, Calling Lake, Alberta.", out.width = '100%'}
knitr::include_graphics("3_output/maps/LiDAR_acqYear_CallingLake.png")
```

##### Calculate LiDAR point count time lag {-}

```{r eval=FALSE}
load("0_data/manual/bird/bird_callingLake_wide.rData")
load("0_data/manual/spatialCov/acqDate_clip.rData")
load("0_data/manual/bird/ss_xy.rData")
acqDate_clip$year<-year(acqDate_clip$date)

ss_xy_timelag<-st_join(ss_xy, acqDate_clip)%>%
  dplyr::rename(c(lidar_year=year,lidar_date=date ))%>%
  dplyr::select(SS, PKEY, lidar_date, lidar_year)

bird_callingLake_wide_2<-left_join(bird_callingLake_wide, ss_xy_timelag)%>%
  dplyr::rename(c(ss_year=year, ss_date=DATE))%>%
  dplyr::select(-geometry)

bird_callingLake_wide_2$SS_lidar_timelag<-bird_callingLake_wide_2$ss_year-bird_callingLake_wide_2$lidar_year

bird_callingLake_wide_2<-bird_callingLake_wide_2[c(1:4, 199:201, 5:198)]
save(bird_callingLake_wide_2, file = "0_data/manual/bird/bird_callingLake_wide_2.rData")
```


##### Filter stations  {-}

Filter stations to those with at least 10 consecutive years of point counts

```{r eval=FALSE}

#Count number of annual surveys done at each station

SS_nYear<-
  bird_callingLake_wide_2%>%
  dplyr::select(SS, ss_year)%>%
  distinct()%>%
  group_by(SS)%>%
  dplyr::summarize(n_years=n())
  save(SS_nYear, file="0_data/manual/bird/SS_nYear.rData")

# add number of of sampling years to bird data frame  
bird_callingLake_wide_3<-left_join(bird_callingLake_wide_2, SS_nYear)
bird_callingLake_wide_3<-bird_callingLake_wide_3[c(1:2, 202, 3:201)]
save(bird_callingLake_wide_3, file = "0_data/manual/bird/bird_callingLake_wide_3.rData")

# keep only bird data where there is more than 10 years of consequtive data

bird_callingLake_wide_4<-
bird_callingLake_wide_3%>%
 filter(n_years>10)  

save(bird_callingLake_wide_4, file = "0_data/manual/bird/bird_callingLake_wide_4.rData")

```













### Select species {-}


#### Calculate number of detections  {-}

```{r eval=FALSE}
load("0_data/manual/bird/bird_callingLake_wide_4.rData")

library(tibble)

#Make a table with total detection for each species. Sort by detection
b<-bird_callingLake_wide_4[-c(1:14)]
totDet<-colSums(b !=0)
totDet<-totDet[order(-totDet)]
totDet<-as.data.frame(totDet)
totDet$percStations<-round(totDet$totDet/nrow(b)*100,2)

# add common species name
bird_alpha<-read.csv("../../data/avian_lifeHistory/IBP-Alpha-Codes20.csv")
ba<-bird_alpha[c(2,4)]
totDet<-left_join(rownames_to_column(totDet), ba, by=c("rowname"="SPEC"))
totDet<-totDet[c(1,4, 2,3)]
names(totDet)[1]<-"spp"

total_species_detections<-totDet
#save(totDet, file ="2_pipeline/store/totDet_qpad.rData")
save(total_species_detections, file="0_data/manual/bird/total_species_detections.rData")

```

```{r echo=FALSE, message = FALSE, results="asis"}
load("0_data/manual/bird/total_species_detections.rData")

knitr::kable(total_species_detections) %>%
  kable_styling(font_size = 10, bootstrap_options= c("striped", "hover", "condensed" ), full_width= F, position="center") %>%
  column_spec(1, width= "20em")%>%
  scroll_box(width = "100%", height = "400px")
```


#### Identify functional traits of species {-}

```{r eval=FALSE}
#Get functional traits from the traitdata package
install.packages("remotes")
remotes::install_github("RS-eco/traitdata")
library(traitdata)

# view glossary of trait data objects
#t<-trait_glossary
el<-elton_birds

load("0_data/manual/bird/bird_callingLake_long_sub.rData")
bird_alpha<-read.csv("../../data/avian_lifeHistory/IBP-Alpha-Codes20.csv")

birds<-bird_callingLake_long_sub%>%
  dplyr::select(SPECIES)%>%
  distinct()
#Get a list of species that don't match alpha codes
birds<-left_join(birds, bird_alpha, by=c("SPECIES"="SPEC"))

library(fuzzyjoin)

bird_traits<-stringdist_left_join(birds, el, by= c("COMMONNAME" = "English"), max_dist=1, ignore_case=TRUE)

bird_traits<-bird_traits[order(bird_traits$SPECIES),]
bird_traits<-bird_traits[-c(2:3)]

save(bird_traits, file="0_data/manual/bird/bird_traits.rData")
```


```{r echo=FALSE, message = FALSE, results="asis"}
load("0_data/manual/bird/bird_traits.rData")

knitr::kable(bird_traits) %>%
  kable_styling(font_size = 10, bootstrap_options= c("striped", "hover", "condensed" ), full_width= F, position="center") %>%
  column_spec(1, width= "20em")%>%
  scroll_box(width = "100%", height = "400px")
```





#### Look for variation in yearly detections {-}

Check to see if the presence of bird species is increasing, decreasing, or staying steady

```{r eval=FALSE}

load("0_data/manual/bird/bird_callingLake_wide_4.rData")

sum_detections_year<-bird_callingLake_wide_4%>%
  group_by(ss_year) %>%
  dplyr::summarise(across(15:201, sum)) %>% #have to subtract the number of grouping columns.
  dplyr::select(order(colnames(.)))%>%
  dplyr::select(ss_year, ALFL:SPSA, SSHA:YRWA)


#standard deviation
sd<-colwise(sd)(sum_detections_year)

#median absolute deviation
mad<-colwise(mad)(sum_detections_year)

#coeficient of variation
ce<-sapply(sum_detections_year, function(x) sd(x) / mean(x))

#variance
var<-colwise(var)(sum_detections_year)

#combine into a single dataframe
bird_detVar<-rbind(sd,mad,ce, var)

#round
bird_detVar<-bird_detVar %>% mutate_if(is.numeric, round, 2)

bird_detVar$stat<-c("standard_deviation", "median_absolute_deviation", "coeficient_of_variation", "variance")

save(bird_detVar, file="0_data/manual/bird/bird_detVar.rData")


bird_ce<-as.data.frame(bird_detVar[3,])
bird_ce<-tidyr::gather(bird_ce, key="species", value="coeficient_of_variation", CONW:BWWA)
bird_ce<-bird_ce[2:3]
bird_ce<-bird_ce[order(bird_ce$species),]
bird_det_ce<-bird_ce
save(bird_det_ce, file="0_data/manual/bird/bird_det_ce.rData")



# a<-ncol(bird_detVar-1)
# 
# bird_detVar<-bird_detVar[c(ncol(bird_detVar), 2:(ncol(bird_detVar)-1) )]
#

```

```{r echo=FALSE, message = FALSE, results="asis"}
load("0_data/manual/bird/bird_det_ce.rData")

knitr::kable(bird_det_ce) %>%
  kable_styling(font_size = 10, bootstrap_options= c("striped", "hover", "condensed" ), full_width= F, position="center") %>%
  column_spec(1, width= "20em")%>%
  scroll_box(width = "100%", height = "400px")
```


join CE and detection rate table for easier viewing

```{r eval=FALSE}
load("0_data/manual/bird/bird_det_ce.rData")
load("0_data/manual/bird/total_species_detections.rData")

bird_totDet_ce<-left_join(total_species_detections, bird_det_ce, by=c("spp"="species"))

save(bird_totDet_ce, file="0_data/manual/bird/bird_totDet_ce.rData")

```

```{r echo=FALSE, message = FALSE, results="asis"}
load("0_data/manual/bird/bird_totDet_ce.rData")

knitr::kable(bird_totDet_ce) %>%
  kable_styling(font_size = 10, bootstrap_options= c("striped", "hover", "condensed" ), full_width= F, position="center") %>%
  column_spec(1, width= "20em")%>%
  scroll_box(width = "100%", height = "400px")
```


```{r eval=FALSE}
bird_totDet_ce_cand<-bird_totDet_ce %>%
  filter(percStations>5 & coeficient_of_variation<.6)

#add  functional traits
load("0_data/manual/bird/bird_traits.rData")
bird_traits2<-bird_traits[c(1, 32:36)]

bird_totDet_ce_cand<-left_join(bird_totDet_ce_cand, bird_traits2, by=c("spp"="SPECIES"))

save(bird_totDet_ce_cand, file="0_data/manual/bird/bird_totDet_ce_cand.rData")
```

```{r echo=FALSE, message = FALSE, results="asis"}
load("0_data/manual/bird/bird_totDet_ce_cand.rData")

knitr::kable(bird_totDet_ce_cand) %>%
  kable_styling(font_size = 10, bootstrap_options= c("striped", "hover", "condensed" ), full_width= F, position="center") %>%
  column_spec(1, width= "20em")%>%
  scroll_box(width = "100%", height = "400px")
```



Based on this, we selected Black-throated Green Warbler (BTNW), associated with mature forests;  Swainson’s Thrush (SWTH) a forest generalist; and  Mourning Warbler (MOWA) an early-seral specialist. 
YBSA Yellow-bellied Sapsucker
AMRE American redstart (mid seral)
WTSP White throated sparrow (early)


```{r eval=FALSE}
load("0_data/manual/bird/bird_callingLake_wide_4.rData")

bird_callingLake_wide_5<-bird_callingLake_wide_4%>%
  dplyr::select(SS:surveyType, AMRE, BTNW, MOWA, SWTH, WIWR, WTSP, YBSA)%>%
  dplyr::rename(c(AMRE_OCC=AMRE, BTNW_OCC=BTNW, MOWA_OCC=MOWA, SWTH_OCC=SWTH, WIWR_OCC=WIWR, WTSP_OCC=WTSP, YBSA_OCC=YBSA))

save(bird_callingLake_wide_5, file="0_data/manual/bird/bird_callingLake_wide_5.rData")
```


```{r echo=FALSE, message = FALSE, results="asis"}
load("0_data/manual/bird/bird_callingLake_wide_5.rData")

knitr::kable(head(bird_callingLake_wide_5)) %>%
  kable_styling(font_size = 10, bootstrap_options= c("striped", "hover", "condensed" ), full_width= F, position="center") %>%
  column_spec(1, width= "20em")%>%
  scroll_box(width = "100%", height = "400px")
```

### Add abundance columns {-}

```{r eval=FALSE}
load("0_data/manual/bird/bird_callingLake_wide_5.rData")
load("0_data/manual/bird/bird_callingLake_wide_abund.rData")

bird_callingLake_wide_abund_2<-bird_callingLake_wide_abund%>%
  dplyr::select(SS:surveyType, AMRE, BTNW, MOWA, SWTH, WIWR, WTSP, YBSA)%>%
   dplyr::rename(c(ss_year=year, ss_date=DATE))%>%
  dplyr::rename(c(AMRE_ABUND=AMRE, BTNW_ABUND=BTNW, MOWA_ABUND=MOWA, SWTH_ABUND=SWTH, WIWR_ABUND=WIWR, WTSP_ABUND=WTSP, YBSA_ABUND=YBSA))

bird_callingLake_wide_6<-left_join(bird_callingLake_wide_5, bird_callingLake_wide_abund_2)
save(bird_callingLake_wide_6, file="0_data/manual/bird/bird_callingLake_wide_6.rData")
```


### Add BAM QPAD offsets {-}

https://github.com/psolymos/QPAD

[@SolymosMatsuoka2013]

```{r eval=FALSE}
load("../../data/avian_lifeHistory/BAM_avianDatabase/AB-data-w-offsets-to-Brendan.RData")

qpad_offsets<-as.data.frame(off)

callingLake_QPAD_offsets <- tibble::rownames_to_column(qpad_offsets, "PKEY")%>%
      dplyr::select(PKEY, AMRE, BTNW, MOWA, SWTH, WIWR, WTSP, YBSA)%>%
      dplyr::rename(c(AMRE_OFF=AMRE, BTNW_OFF=BTNW, MOWA_OFF=MOWA, SWTH_OFF=SWTH, WIWR_OFF= WIWR, WTSP_OFF=WTSP, YBSA_OFF=YBSA))%>%
      dplyr::filter(str_detect(PKEY, "^CL"))

save(callingLake_QPAD_offsets, file="0_data/manual/bird/callingLake_QPAD_offsets.rData")
```


```{r echo=FALSE, message = FALSE, results="asis"}
load("0_data/manual/bird/callingLake_QPAD_offsets.rData")
6
knitr::kable(head(callingLake_QPAD_offsets)) %>%
  kable_styling(font_size = 10, bootstrap_options= c("striped", "hover", "condensed" ), full_width= F, position="center") %>%
  column_spec(1, width= "20em")%>%
  scroll_box(width = "100%", height = "400px")
```
