==> 1_code/rNotebooks/1_setup/4_combine_data.Rmd <==
```{r setup, include=FALSE, cache=FALSE}
#Set root directory to R project root
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
```



```{css, echo=FALSE}
# set max height of code chunks using the following css styling

pre {
  max-height: 300px;
  overflow-y: auto;
}

pre[class] {
  max-height: 500px;
}
```

```{r eval=FALSE}
library(dplyr)
```



### Combine spatial data {-}

Combine spatial data into a single data frame

#### Load data {-}

```{r eval=FALSE}
# bird detection data
load("0_data/manual/bird/bird_callingLake_wide_6.rData")

# QPAD offsets
load("0_data/manual/bird/callingLake_QPAD_offsets.rData")

# LIDAR
load("0_data/manual/spatialCov/LiDAR/ss_lidar_df.rData")

# Classified landcover
load("0_data/manual/spatialCov/CASFRI/pkey_cas_cov.rData")
load("0_data/manual/spatialCov/QPAD/callingLake_QPAD_LC.rData")

# NDVI
load("0_data/manual/spatialCov/ndvi/ss_ndvi_df.rData")

```


#### Select covariates {-}

- **CAS-FRI derived**
  - rast_forest_age
  - ss_forest_origDate_mean
  - ss_forest_origDate_sd
  - Forest_Type 
  - poly_forest_age

- **From 2005 MODIS Landcover**
  - LCC2
  - LCC4
  
- **LiDAR**  
  - canopy_relief_ratio                      
  - elev_0pnt15_to_2pnt00_return_proportion  
  - elev_10pnt00_to_15pnt00_return_proportion
  - elev_15pnt00_to_20pnt00_return_proportion
  - elev_20pnt00_to_25pnt00_return_proportion
  - elev_25pnt00_to_30pnt00_return_proportion
  - elev_2pnt00_to_4pnt00_return_proportion  
  - elev_30pnt00_to_50pnt00_return_proportion
  - elev_4pnt00_to_6pnt00_return_proportion  
  - elev_6pnt00_to_8pnt00_return_proportion  
  - elev_8pnt00_to_10pnt00_return_proportion 
  - elev_below_0pnt15_return_proportion      
  - elev_cv                                  
  - elev_kurtosis                            
  - elev_maximum                             
  - elev_mean                                
  - elev_p50                                 
  - elev_p95                                 
  - elev_p99                                 
  - elev_stddev                              
  - percentage_first_returns_above_2pnt00    
  - percentage_first_returns_above_mean      
  - total_all_returns
  
- **LiDAR**  
  - ndvi_1990
  - ndvi_1991
  - ndvi_1992
  - ndvi_1993
  - ndvi_1994
  - ndvi_1995
  - ndvi_1996
  - ndvi_1997
  - ndvi_1998
  - ndvi_1999
  - ndvi_2000
  - ndvi_2001
  - ndvi_2002
  - ndvi_2003
  - ndvi_2004
  - ndvi_2005
  - ndvi_2006
  - ndvi_2007
  - ndvi_2008
  - ndvi_2009
  - ndvi_2010
  - ndvi_2011
  - ndvi_2012
  - ndvi_2013
  - ndvi_2014
  - ndvi_2015
  - ndvi_2016
  - ndvi_2017
  - ndvi_2018
  - ndvi_2019
  - ndvi_2020
  

```{r eval=FALSE}
# LiDAR
l<-ss_lidar_df %>%
  dplyr::select(SS, canopy_relief_ratio, elev_0pnt15_to_2pnt00_return_proportion, elev_10pnt00_to_15pnt00_return_proportion, elev_15pnt00_to_20pnt00_return_proportion, elev_20pnt00_to_25pnt00_return_proportion, elev_25pnt00_to_30pnt00_return_proportion, elev_2pnt00_to_4pnt00_return_proportion, elev_30pnt00_to_50pnt00_return_proportion, elev_4pnt00_to_6pnt00_return_proportion, elev_6pnt00_to_8pnt00_return_proportion, elev_8pnt00_to_10pnt00_return_proportion , elev_below_0pnt15_return_proportion, elev_cv, elev_kurtosis, elev_maximum , elev_mean, elev_p05 , elev_p10 , elev_p20 , elev_p25 , elev_p30 , elev_p50 , elev_p60 , elev_p70 , elev_p75 , elev_p80 , elev_p95 , elev_p99 , elev_stddev, percentage_first_returns_above_2pnt00, percentage_first_returns_above_mean, total_all_returns)

# CAS-FRI
c<-pkey_cas_cov %>%
  dplyr::select(SS, PKEY, rast_forest_age, forest_age_class_r, ss_forest_origDate_mean, ss_forest_origDate_sd, Forest_Type, poly_forest_age, forest_age_class_p, 
)

# MODIS
m<-callingLake_QPAD_LC %>%
  dplyr::select(SS, PKEY, LCC2, LCC4)

# Landsat NDVI
n<-ss_ndvi_df
```


#### Join spatial data {-}

```{r eval=FALSE}
spatialCov_all<-left_join(c, m)%>%
  left_join(l)%>%
  left_join()

save(spatialCov_all, file = "2_pipeline/store/spatialCov_all.rData")
```



### Combine observation and spatial data {-}


#### Join data {-}

```{r eval=FALSE}

bd<-distinct(bird_callingLake_wide_6)

data_main<-bd%>%
  dplyr::select(-lidar_date)%>%
  dplyr::rename(n_survey_years=n_years) %>%
  left_join(callingLake_QPAD_offsets) %>% 
  left_join(spatialCov_all, by="PKEY") %>%
  dplyr::select(-c(SS.y)) %>%
  dplyr::rename(SS=SS.x) %>%
  distinct()

save(data_main, file = "0_data/manual/data_main.rData")
```

```{r echo=FALSE, message = FALSE, results="asis"}
load("0_data/manual/data_main.rData")

knitr::kable(head(data_main)) %>%
  kable_styling(font_size = 10, bootstrap_options= c("striped", "hover", "condensed" ), full_width= F, position="center") %>%
  column_spec(1, width= "20em")%>%
  scroll_box(width = "100%", height = "400px")
```



#### Lag NDVI {-}

```{r eval=FALSE}
l_08<-data_main%>%
  filter(lidar_year=="2008")

l_08$ndvi_lag_0<-l_08$ndvi_2008
l_08$ndvi_lag_1<-l_08$ndvi_2007
l_08$ndvi_lag_2<-l_08$ndvi_2006
l_08$ndvi_lag_3<-l_08$ndvi_2005
l_08$ndvi_lag_4<-l_08$ndvi_2004
l_08$ndvi_lag_5<-l_08$ndvi_2003
l_08$ndvi_lag_6<-l_08$ndvi_2002
l_08$ndvi_lag_7<-l_08$ndvi_2001
l_08$ndvi_lag_8<-l_08$ndvi_2000
l_08$ndvi_lag_9<-l_08$ndvi_1999
l_08$ndvi_lag_10<-l_08$ndvi_1998
l_08$ndvi_lag_11<-l_08$ndvi_1997
l_08$ndvi_lag_12<-l_08$ndvi_1996
l_08$ndvi_lag_13<-l_08$ndvi_1995
l_08$ndvi_lag_14<-l_08$ndvi_1994
l_08$ndvi_lag_15<-l_08$ndvi_1993
l_08$ndvi_lag_16<-l_08$ndvi_1992



l_09<-data_main%>%
  filter(lidar_year=="2009")

l_09$ndvi_lag_0<-l_09$ndvi_2009
l_09$ndvi_lag_1<-l_09$ndvi_2008
l_09$ndvi_lag_2<-l_09$ndvi_2007
l_09$ndvi_lag_3<-l_09$ndvi_2006
l_09$ndvi_lag_4<-l_09$ndvi_2005
l_09$ndvi_lag_5<-l_09$ndvi_2004
l_09$ndvi_lag_6<-l_09$ndvi_2003
l_09$ndvi_lag_7<-l_09$ndvi_2002
l_09$ndvi_lag_8<-l_09$ndvi_2001
l_09$ndvi_lag_9<-l_09$ndvi_2000
l_09$ndvi_lag_10<-l_09$ndvi_1999
l_09$ndvi_lag_11<-l_09$ndvi_1998
l_09$ndvi_lag_12<-l_09$ndvi_1997
l_09$ndvi_lag_13<-l_09$ndvi_1996
l_09$ndvi_lag_14<-l_09$ndvi_1995
l_09$ndvi_lag_15<-l_09$ndvi_1994
l_09$ndvi_lag_16<-l_09$ndvi_1993

data_main_2<-rbind(l_08, l_09)

save(data_main_2, file = "0_data/manual/data_main_2.rData")
```



#### Combine raster data into a single raster stack


```{r eval=FALSE}
#import raster stacks
#LiDAR
lidar_stack<-stackOpen("0_data/manual/spatialCov/LiDAR/lidar_stack.stk")
# scaled_lidar_stack<-scale(lidar_stack)

#NDVI
ndvi_stack<-stackOpen("0_data/external/CL_Landsat_NDVI/ndvi_stack.stk")
#ndvi_stack_pr<-projectRaster(ndvi_stack, crs=lidar_stack)

# scaled_ndvi_stack<-scale(ndvi_stack)

#Forest Age
Age_stack<-stackOpen("0_data/manual/spatialCov/CASFRI/AgeRasters/Age_stack.stk")
Age_stack_pr<-projectRaster(Age_stack, crs=lidar_stack)

#age class
ageClass_stack<-stackOpen("0_data/manual/spatialCov/CASFRI/ageClassRasters/ageClass_stack.stk")
ageClass_stack_pr<-projectRaster(ageClass_stack, crs=lidar_stack)


#study area
load("0_data/manual/bird/studyarea_big.rData")
 
# Crop rasters to study area
ndvi_stack_c<- raster::crop(ndvi_stack, as_Spatial(st_geometry(c_bb))) 
lidar_stack_c<- raster::crop(lidar_stack, as_Spatial(st_geometry(c_bb)))
Age_stack_pr_c<- raster::crop(Age_stack_pr, as_Spatial(st_geometry(c_bb))) 
ageClass_stack_pr_c<- raster::crop(ageClass_stack_pr, as_Spatial(st_geometry(c_bb))) 

                             
#resample so they have the same resolution
ndvi_stack_c_rs<-resample(ndvi_stack_c, ageClass_stack_pr_c)
Age_stack_c_rs<-resample(Age_stack_pr_c, ageClass_stack_pr_c)
lidar_stack_c_rs<-resample(lidar_stack_c, ageClass_stack_pr_c)

lidar_stack_c_rs<-lidar_stack_c_c_rs
# ndvi_stack_c_rs<-resample(ndvi_stack_c, ageClass_stack)
# Age_stack_c_rs<-resample(Age_stack_pr, lidar_stack_c)
# ageClass_stack_pr_c_rs<-resample(ageClass_stack_pr_c, lidar_stack_c)

#rename ndvi raster names to remove "cl_" prefix
names(ndvi_stack_c_rs) <- substring(names(ndvi_stack_c_rs), 4)

# Stack covariates
cov_stack<-stack(lidar_stack_c_rs, ndvi_stack_c_rs, Age_stack_c_rs, ageClass_stack_pr_c)

##### save transformed rasters
#first unstack
d1<-cov_stack
d2 <- unstack(d1)
# save seperate rasters
outputnames <- paste("0_data/manual/spatialCov/transformed_and_resampled/", names(d1), ".tif", sep="")
for(i in seq_along(d2)){writeRaster(d2[[i]], file=outputnames[i], overwrite=TRUE)}

# restack
#list the files
cov_list<-list.files(path="0_data/manual/spatialCov/transformed_and_resampled", pattern = "*.tif$", full.names = TRUE)
cov_stack2<-raster::stack(cov_list)

stackSave(cov_stack2,file="0_data/manual/spatialCov/transformed_and_resampled/cov_stack2.stk")


################################
##Scale rasters for predictions bsed on scaled covariates
################################

scaled_stack<-scale(subset(cov_stack2, c(1:23, 47:111)))

#add back the age class rasters
scaled_stack1<-stack(scaled_stack, subset(cov_stack2, 24:46))

##### save transformed rasters
#first unstack
d3<-scaled_stack1
d4 <- unstack(d3)
# save seperate rasters
outputnames <- paste("0_data/manual/spatialCov/scaled_cov/", names(d3), ".tif", sep="")
for(i in seq_along(d4)){writeRaster(d4[[i]], file=outputnames[i], overwrite=TRUE)}

# restack
#list the files
scaled_cov_list<-list.files(path="0_data/manual/spatialCov/scaled_cov", pattern = "*.tif$", full.names = TRUE)
scaled_stack2<-raster::stack(scaled_cov_list)
stackSave(scaled_stack2,file="0_data/manual/spatialCov/scaled_cov/scaled_stack2.stk")

scaled_cov<-stackOpen("0_data/manual/spatialCov/scaled_cov/scaled_stack2.stk")


cov_stack2<-stackOpen("0_data/manual/spatialCov/transformed_and_resampled/cov_stack2.stk")

```
