==> 1_code/rNotebooks/1_setup/5_format_and_filter.Rmd <==
```{r setup, include=FALSE, cache=FALSE}
#Set root directory to R project root
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
```



```{css, echo=FALSE}
# set max height of code chunks using the following css styling

pre {
  max-height: 300px;
  overflow-y: auto;
}

pre[class] {
  max-height: 500px;
}
```

```{r eval=FALSE}
library(dplyr)
```

### Load data {-}

```{r}
load("0_data/manual/data_main.rData")

#data_main<-data_main%>%dplyr::select(-poly_forest_age)
```


### Filter data {-}

Filter by:
- To minimize the influence of forest edges and adjacent differently aged forest, remove stations that are surrounded by a variety of forest ages (sd>5).


```{r eval=FALSE}
f1<-data_main%>%
  filter(ss_forest_origDate_sd<10)

nrow(f1 %>%
  dplyr::select("SS")%>%
  distinct())


f2<-data_main%>%
  filter(ss_forest_origDate_sd<5)

nrow(f2 %>%
  dplyr::select("SS")%>%
  distinct())

f3a<-data_main %>% 
  filter(rast_forest_age > 30 & ss_forest_origDate_sd<11)
f3b<-data_main %>% 
  filter(rast_forest_age < 31 & ss_forest_origDate_sd<6)  
f3c<-rbind(f3a, f3b)


f4a<-data_main %>%
  filter(rast_forest_age > 33 & ss_forest_origDate_sd<11)
f4b<-data_main %>%
  filter(rast_forest_age < 34 & ss_forest_origDate_sd<6)
f4c<-rbind(f4a, f4b)


data_main_filter<-f4c

save(data_main_filter, file = "2_pipeline/store/data_main_filter.rData")
```

### Reclassify data {-}

```{r eval=FALSE}
d<-data_main_filter
#d<-data_main

# check classes
lapply(d,class)


#Convert character variables to factors
d2 <- cbind(d[1:2],as.data.frame(unclass(d[c(3:ncol(d))]),                 
                       stringsAsFactors = TRUE))
```

### Remove missing data {-}

```{r eval=FALSE}
d3<-d2[complete.cases(d2),]
```

### Scale metrics {-}

```{r eval=FALSE}

# scale Lidar
d3$canopy_relief_ratio<-scale(d3$canopy_relief_ratio, center=TRUE, scale=TRUE)
d3$elev_0pnt15_to_2pnt00_return_proportion<-scale(d3$elev_0pnt15_to_2pnt00_return_proportion, center=TRUE, scale=TRUE)
d3$elev_10pnt00_to_15pnt00_return_proportion<-scale(d3$elev_10pnt00_to_15pnt00_return_proportion, center=TRUE, scale=TRUE)
d3$elev_15pnt00_to_20pnt00_return_proportion<-scale(d3$elev_15pnt00_to_20pnt00_return_proportion, center=TRUE, scale=TRUE)
d3$elev_20pnt00_to_25pnt00_return_proportion<-scale(d3$elev_20pnt00_to_25pnt00_return_proportion, center=TRUE, scale=TRUE)
d3$elev_25pnt00_to_30pnt00_return_proportion<-scale(d3$elev_25pnt00_to_30pnt00_return_proportion, center=TRUE, scale=TRUE)
d3$elev_2pnt00_to_4pnt00_return_proportion<-scale(d3$elev_2pnt00_to_4pnt00_return_proportion, center=TRUE, scale=TRUE)
d3$elev_30pnt00_to_50pnt00_return_proportion<-scale(d3$elev_30pnt00_to_50pnt00_return_proportion, center=TRUE, scale=TRUE)
d3$elev_4pnt00_to_6pnt00_return_proportion<-scale(d3$elev_4pnt00_to_6pnt00_return_proportion, center=TRUE, scale=TRUE)
d3$elev_6pnt00_to_8pnt00_return_proportion<-scale(d3$elev_6pnt00_to_8pnt00_return_proportion, center=TRUE, scale=TRUE)
d3$elev_8pnt00_to_10pnt00_return_proportion <-scale(d3$elev_8pnt00_to_10pnt00_return_proportion , center=TRUE, scale=TRUE)
d3$elev_below_0pnt15_return_proportion<-scale(d3$elev_below_0pnt15_return_proportion, center=TRUE, scale=TRUE)
d3$elev_cv<-scale(d3$elev_cv, center=TRUE, scale=TRUE)
d3$elev_kurtosis<-scale(d3$elev_kurtosis, center=TRUE, scale=TRUE)
d3$elev_maximum <-scale(d3$elev_maximum , center=TRUE, scale=TRUE)
d3$elev_mean<-scale(d3$elev_mean, center=TRUE, scale=TRUE)
d3$elev_p05 <-scale(d3$elev_p05 , center=TRUE, scale=TRUE)
d3$elev_p10 <-scale(d3$elev_p10 , center=TRUE, scale=TRUE)
d3$elev_p20 <-scale(d3$elev_p20 , center=TRUE, scale=TRUE)
d3$elev_p25 <-scale(d3$elev_p25 , center=TRUE, scale=TRUE)
d3$elev_p30 <-scale(d3$elev_p30 , center=TRUE, scale=TRUE)
d3$elev_p50 <-scale(d3$elev_p50 , center=TRUE, scale=TRUE)
d3$elev_p60 <-scale(d3$elev_p60 , center=TRUE, scale=TRUE)
d3$elev_p70 <-scale(d3$elev_p70 , center=TRUE, scale=TRUE)
d3$elev_p75 <-scale(d3$elev_p75 , center=TRUE, scale=TRUE)
d3$elev_p80 <-scale(d3$elev_p80 , center=TRUE, scale=TRUE)
d3$elev_p95 <-scale(d3$elev_p95 , center=TRUE, scale=TRUE)
d3$elev_p99 <-scale(d3$elev_p99 , center=TRUE, scale=TRUE)
d3$elev_stddev<-scale(d3$elev_stddev, center=TRUE, scale=TRUE)
d3$percentage_first_returns_above_2pnt00<-scale(d3$percentage_first_returns_above_2pnt00, center=TRUE, scale=TRUE)
d3$percentage_first_returns_above_mean<-scale(d3$percentage_first_returns_above_mean, center=TRUE, scale=TRUE)
d3$total_all_returns <-scale(d3$total_all_returns , center=TRUE, scale=TRUE)

d3$rast_forest_age<-scale(d3$rast_forest_age, center=TRUE, scale=TRUE)
#d3$poly_forest_age<-scale(d3$poly_forest_age, center=TRUE, scale=TRUE)

d3$ndvi_1990<-scale(d3$ndvi_1990, center=TRUE, scale=TRUE)
d3$ndvi_1991<-scale(d3$ndvi_1991, center=TRUE, scale=TRUE)
d3$ndvi_1992<-scale(d3$ndvi_1992, center=TRUE, scale=TRUE)
d3$ndvi_1993<-scale(d3$ndvi_1993, center=TRUE, scale=TRUE)
d3$ndvi_1994<-scale(d3$ndvi_1994, center=TRUE, scale=TRUE)
d3$ndvi_1995<-scale(d3$ndvi_1995, center=TRUE, scale=TRUE)
d3$ndvi_1996<-scale(d3$ndvi_1996, center=TRUE, scale=TRUE)
d3$ndvi_1997<-scale(d3$ndvi_1997, center=TRUE, scale=TRUE)
d3$ndvi_1998<-scale(d3$ndvi_1998, center=TRUE, scale=TRUE)
d3$ndvi_1999<-scale(d3$ndvi_1999, center=TRUE, scale=TRUE)
d3$ndvi_2000<-scale(d3$ndvi_2000, center=TRUE, scale=TRUE)
d3$ndvi_2001<-scale(d3$ndvi_2001, center=TRUE, scale=TRUE)
d3$ndvi_2002<-scale(d3$ndvi_2002, center=TRUE, scale=TRUE)
d3$ndvi_2003<-scale(d3$ndvi_2003, center=TRUE, scale=TRUE)
d3$ndvi_2004<-scale(d3$ndvi_2004, center=TRUE, scale=TRUE)
d3$ndvi_2005<-scale(d3$ndvi_2005, center=TRUE, scale=TRUE)
d3$ndvi_2006<-scale(d3$ndvi_2006, center=TRUE, scale=TRUE)
d3$ndvi_2007<-scale(d3$ndvi_2007, center=TRUE, scale=TRUE)
d3$ndvi_2008<-scale(d3$ndvi_2008, center=TRUE, scale=TRUE)
d3$ndvi_2009<-scale(d3$ndvi_2009, center=TRUE, scale=TRUE)
d3$ndvi_2010<-scale(d3$ndvi_2010, center=TRUE, scale=TRUE)
d3$ndvi_2011<-scale(d3$ndvi_2011, center=TRUE, scale=TRUE)
d3$ndvi_2012<-scale(d3$ndvi_2012, center=TRUE, scale=TRUE)
d3$ndvi_2013<-scale(d3$ndvi_2013, center=TRUE, scale=TRUE)
d3$ndvi_2014<-scale(d3$ndvi_2014, center=TRUE, scale=TRUE)
d3$ndvi_2015<-scale(d3$ndvi_2015, center=TRUE, scale=TRUE)
d3$ndvi_2016<-scale(d3$ndvi_2016, center=TRUE, scale=TRUE)
d3$ndvi_2017<-scale(d3$ndvi_2017, center=TRUE, scale=TRUE)
d3$ndvi_2018<-scale(d3$ndvi_2018, center=TRUE, scale=TRUE)
d3$ndvi_2019<-scale(d3$ndvi_2019, center=TRUE, scale=TRUE)
d3$ndvi_2020<-scale(d3$ndvi_2020, center=TRUE, scale=TRUE)


sdm_d3<-d3
save(sdm_d3, file="2_pipeline/tmp/sdm_d3.rData")

```


***

### Group by time since LiDAR {-}


#### Identify stations that have surveys at each time step

```{r eval=FALSE}

dd<-d3
dd<-dd%>%
  filter(SS_lidar_timelag < "1")%>%
  filter(SS_lidar_timelag>"-16")

complete_series_SS<-as.data.frame(Reduce(intersect, split(dd$SS, dd$SS_lidar_timelag)))%>%
  setNames("SS")

save(complete_series_SS, file="2_pipeline/tmp/complete_series_SS.rData")
```

#### Filter dataframe to include only those stations

```{r eval=FALSE}
d4<-semi_join(d3, complete_series_SS)
sdm_d<-d4
save(sdm_d, file="2_pipeline/tmp/sdm_d.rData")
```

#### Seperate by aqusition year

```{r eval=FALSE}
d4<-sdm_d
d5_2008<-filter(d4, lidar_year==2008)
d5_2009<-filter(d4, lidar_year==2009)
```

#### Check variability in detections by year
```{r eval=FALSE}
# Median and Median Absolute Deviation (MAD)


#x<-d5_2009%>%
x<-d4%>%
  filter(SS_lidar_timelag<1)%>%
  group_by(SS_lidar_timelag)%>%
  #filter(lidar_year==2009)%>%
  dplyr::summarise(across(13:15, sum)) #


   # ce<-sapply(test[2:4], function(x) sd(x) / mean(x))
   # mean<-sapply(test, function(x) mean(x))
   # dist<-sapply(test, function(x) x - mean(x))
   
  
xBTNW<-x$BTNW
med = median(xBTNW)
# subtract median from each value of x and get absolute deviation
abs_dev = abs(xBTNW-med)
# get MAD
mad = 1.4826 * median(abs_dev)

# get threshold values for outliers
Tmin = med-(1*mad) 
Tmax = med+(1*mad) 

# find outlier
xBTNW[which(xBTNW < Tmin | xBTNW > Tmax)]

# remove outlier
yBTNW=xBTNW[which(xBTNW > Tmin & xBTNW < Tmax)]


xMOWA<-x$MOWA
med = median(xMOWA)
# subtract median from each value of x and get absolute deviation
abs_dev = abs(xMOWA-med)
# get MAD
mad = 1.4826 * median(abs_dev)

# get threshold values for outliers
Tmin = med-(.5*mad) 
Tmax = med+(.5*mad) 

# find outlier
xMOWA[which(xMOWA < Tmin | xMOWA > Tmax)]

# remove outlier
yMOWA=xMOWA[which(xMOWA > Tmin & xMOWA < Tmax)]


xSWTH<-x$SWTH
med = median(xSWTH)
# subtract median from each value of x and get absolute deviation
abs_dev = abs(xSWTH-med)
# get MAD
mad = 1.4826 * median(abs_dev)

# get threshold values for outliers
Tmin = med-(2*mad) 
Tmax = med+(2*mad) 

# find outlier
xSWTH[which(xSWTH < Tmin | xSWTH > Tmax)]

# remove outlier
ySWTH=xSWTH[which(xSWTH > Tmin & xSWTH < Tmax)]


y2=x %>%
  filter(BTNW %in% yBTNW)#%>%
  #filter(MOWA%in% yMOWA)#%>%
  #filter(SWTH %in% ySWTH)



ce<-sapply(y2, function(x) sd(x) / mean(x))
```




```{r eval=FALSE, echo=FALSE, include=FALSE}

# d_minus16<-sdm_d%>%
#   filter(SS_lidar_timelag==-16)
# nrow(distinct(as.data.frame((d_minus16$SS))))
# d_minus16_ss<-d_minus15%>%dplyr::select(SS, SS_lidar_timelag)%>%distinct()
# 
# d_minus15<-sdm_d%>%
#   filter(SS_lidar_timelag==-15)
# nrow(distinct(as.data.frame((d_minus15$SS))))
# d_minus15_ss<-d_minus15%>%dplyr::select(SS, SS_lidar_timelag)%>%distinct()
# 
# d_minus14<-sdm_d%>%
#   filter(SS_lidar_timelag==-14)
# nrow(distinct(as.data.frame((d_minus14$SS))))
# d_minus14_ss<-d_minus14%>%dplyr::select(SS, SS_lidar_timelag)%>%distinct()
# 
# d_minus13<-sdm_d%>%
#   filter(SS_lidar_timelag==-13)
# nrow(distinct(as.data.frame((d_minus13$SS))))
# d_minus13_ss<-d_minus13%>%dplyr::select(SS, SS_lidar_timelag)%>%distinct()
# 
# d_minus12<-sdm_d%>%
#   filter(SS_lidar_timelag==-12)
# nrow(distinct(as.data.frame((d_minus12$SS))))
# d_minus12_ss<-d_minus12%>%dplyr::select(SS, SS_lidar_timelag)%>%distinct()
# 
# d_minus11<-sdm_d%>%
#   filter(SS_lidar_timelag==-11)
# nrow(distinct(as.data.frame((d_minus11$SS))))
# d_minus11_ss<-d_minus11%>%dplyr::select(SS, SS_lidar_timelag)%>%distinct()
# 
# d_minus10<-sdm_d%>%
#   filter(SS_lidar_timelag==-10)
# nrow(distinct(as.data.frame((d_minus10$SS))))
# d_minus10_ss<-d_minus10%>%dplyr::select(SS, SS_lidar_timelag)%>%distinct()
# 
# d_minus9<-sdm_d%>%
#   filter(SS_lidar_timelag==-9)
# nrow(distinct(as.data.frame((d_minus9$SS))))
# d_minus9_ss<-d_minus9%>%dplyr::select(SS, SS_lidar_timelag)%>%distinct()
# 
# d_minus8<-sdm_d%>%
#   filter(SS_lidar_timelag==-8)
# nrow(distinct(as.data.frame((d_minus8$SS))))
# d_minus8_ss<-d_minus8%>%dplyr::select(SS, SS_lidar_timelag)%>%distinct()
# 
# d_minus7<-sdm_d%>%
#   filter(SS_lidar_timelag==-7)
# nrow(distinct(as.data.frame((d_minus7$SS))))
# d_minus7_ss<-d_minus7%>%dplyr::select(SS, SS_lidar_timelag)%>%distinct()
# 
# d_minus6<-sdm_d%>%
#   filter(SS_lidar_timelag==-6)
# nrow(distinct(as.data.frame((d_minus6$SS))))
# d_minus6_ss<-d_minus6%>%dplyr::select(SS, SS_lidar_timelag)%>%distinct()
# 
# d_minus5<-sdm_d%>%
#   filter(SS_lidar_timelag==-5)
# nrow(distinct(as.data.frame((d_minus5$SS))))
# d_minus5_ss<-d_minus5%>%dplyr::select(SS, SS_lidar_timelag)%>%distinct()
# 
# d_minus4<-sdm_d%>%
#   filter(SS_lidar_timelag==-4)
# nrow(distinct(as.data.frame((d_minus4$SS))))
# d_minus4_ss<-d_minus4%>%dplyr::select(SS, SS_lidar_timelag)%>%distinct()
# 
# d_minus3<-sdm_d%>%
#   filter(SS_lidar_timelag==-3)
# nrow(distinct(as.data.frame((d_minus3$SS))))
# d_minus3_ss<-d_minus3%>%dplyr::select(SS, SS_lidar_timelag)%>%distinct()
# 
# d_minus2<-sdm_d%>%
#   filter(SS_lidar_timelag==-2)
# nrow(distinct(as.data.frame((d_minus2$SS))))
# d_minus2_ss<-d_minus2%>%dplyr::select(SS, SS_lidar_timelag)%>%distinct()
# 
# d_minus1<-sdm_d%>%
#   filter(SS_lidar_timelag==-1)
# nrow(distinct(as.data.frame((d_minus1$SS))))
# d_minus1_ss<-d_minus1%>%dplyr::select(SS, SS_lidar_timelag)%>%distinct()
# 
# d_0<-sdm_d%>%
#   filter(SS_lidar_timelag==0)
# nrow(distinct(as.data.frame((d_0$SS))))
# d_0_ss<-d_0%>%dplyr::select(SS, SS_lidar_timelag)%>%distinct()
# 
# 
# d_1<-sdm_d%>%
#   filter(SS_lidar_timelag==1)
# nrow(distinct(as.data.frame((d_1$SS))))
# d_1_ss<-d_1%>%dplyr::select(SS, SS_lidar_timelag)%>%distinct()
# 
# d_2<-sdm_d%>%
#   filter(SS_lidar_timelag==2)
# nrow(distinct(as.data.frame((d_2$SS))))
# d_2_ss<-d_2%>%dplyr::select(SS, SS_lidar_timelag)%>%distinct()
# 
# d_3<-sdm_d%>%
#   filter(SS_lidar_timelag==3)
# nrow(distinct(as.data.frame((d_3$SS))))
# d_3_ss<-d_3%>%dplyr::select(SS, SS_lidar_timelag)%>%distinct()
# 
# d_4<-sdm_d%>%
#   filter(SS_lidar_timelag==4)
# nrow(distinct(as.data.frame((d_4$SS))))
# d_4_ss<-d_4%>%dplyr::select(SS, SS_lidar_timelag)%>%distinct()
# 
# d_5<-sdm_d%>%
#   filter(SS_lidar_timelag==5)
# nrow(distinct(as.data.frame((d_5$SS))))
# d_5_ss<-d_5%>%dplyr::select(SS, SS_lidar_timelag)%>%distinct()
# 
# d_6<-sdm_d%>%
#   filter(SS_lidar_timelag==6)
# nrow(distinct(as.data.frame((d_6$SS))))
# d_6_ss<-d_6%>%dplyr::select(SS, SS_lidar_timelag)%>%distinct()

```
